eliminated a large number of warnings, many of which had to do with mixing types.  i switched some types to size_t and a couple to ptrdiff_t to make things more consistent, but there is a huge amount of work to be done to make types consistent.  int, unsigned int, unsigned, off_t, and other types are ill-suited to storing buffer lengths or memory addresses.

+++ libclamav/pdf.h
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2015 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
+ *  Copyright (C) 2015, 2017 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
  *  Copyright (C) 2007-2013 Sourcefire, Inc.
  *
  *  Authors: Nigel Horne
@@ -160,7 +160,7 @@ int pdf_findobj(struct pdf_struct *pdf);
 struct pdf_obj *find_obj(struct pdf_struct *pdf, struct pdf_obj *obj, uint32_t objid);
 
 void pdf_handle_enc(struct pdf_struct *pdf);
-char *decrypt_any(struct pdf_struct *pdf, uint32_t id, const char *in, off_t *length, enum enc_method enc_method);
+char *decrypt_any(struct pdf_struct *pdf, uint32_t id, const char *in, size_t *length, enum enc_method enc_method);
 enum enc_method get_enc_method(struct pdf_struct *pdf, struct pdf_obj *obj);
 enum enc_method parse_enc_method(const char *dict, unsigned len, const char *key, enum enc_method def);
